Download all books on a specific topic into the same directory
GPT output:
In this version of the function, topic is used to create the directory, and all files will be downloaded into this directory. The directory name is also cleaned of any non-alphanumeric and non-space characters, as these could cause issues on some filesystems.

Remember that you would also need to update wherever this function is called in the code, to pass the topic argument.